ddl: support multi-schema-change

+++ ddl/ddl_ops.go
@@ -79,6 +79,9 @@ func (c *testCase) generateDDLOps() error {
 	if err := c.generateModifyColumn2(5); err != nil {
 		return errors.Trace(err)
 	}
+	if err := c.generateMultiSchemaChange(defaultTime); err != nil {
+		return errors.Trace(err)
+	}
 	return nil
 }
 
@@ -108,6 +111,8 @@ const (
 	// ddlModifyColumn2 is used to test column type change.
 	ddlModifyColumn2
 
+	ddlMultiSchemaChange
+
 	ddlKindNil
 )
 
@@ -135,6 +140,8 @@ var mapOfDDLKind = map[string]DDLKind{
 
 	"modify column":  ddlModifyColumn,
 	"modify column2": ddlModifyColumn2,
+
+	"multi schema change": ddlMultiSchemaChange,
 }
 
 var mapOfDDLKindToString = map[DDLKind]string{
@@ -160,6 +167,7 @@ var mapOfDDLKindToString = map[DDLKind]string{
 	ddlModifyTableCharsetAndCollate: "modify table charset and collate",
 	ddlModifyColumn:                 "modify column",
 	ddlModifyColumn2:                "modify column2",
+	ddlMultiSchemaChange:            "multi schema change",
 }
 
 // mapOfDDLKindProbability use to control every kind of ddl request execute probability.
@@ -174,6 +182,8 @@ var mapOfDDLKindProbability = map[DDLKind]float64{
 	ddlModifyColumn:  0.5,
 	ddlModifyColumn2: 0.5,
 	ddlDropColumn:    0.5,
+
+	ddlMultiSchemaChange: 0.5,
 
 	ddlCreateView: 0.30,
 
@@ -253,6 +263,8 @@ func (c *testCase) updateTableInfo(task *ddlJobTask) error {
 		return c.setDefaultValueJob(task)
 	case ddlModifyColumn2:
 		return c.modifyColumnJob(task)
+	case ddlMultiSchemaChange:
+		return c.multiSchemaChangeJob(task)
 	}
 	return fmt.Errorf("unknow ddl task , %v", *task)
 }
@@ -457,6 +469,14 @@ func getLastDDLInfo(conn *sql.Conn) (uint64, string, error) {
 		return 0, "", err
 	}
 	return seqNum, query, row.Close()
+}
+
+func (c *testCase) getTable(t interface{}) *ddlTestTable {
+	if t == nil {
+		return c.pickupRandomTable()
+	} else {
+		return t.(*multiSchemaChangeCtx).tblInfo
+	}
 }
 
 /*
@@ -1143,8 +1163,8 @@ func generateIndexSignture(index ddlTestIndex) string {
 	return signature
 }
 
-func (c *testCase) prepareAddIndex(_ interface{}, taskCh chan *ddlJobTask) error {
-	table := c.pickupRandomTable()
+func (c *testCase) prepareAddIndex(ctx interface{}, taskCh chan *ddlJobTask) error {
+	table := c.getTable(ctx)
 	if table == nil {
 		return nil
 	}
@@ -1187,6 +1207,12 @@ func (c *testCase) prepareAddIndex(_ interface{}, taskCh chan *ddlJobTask) error
 			if column.canBeIndex() {
 				index.columns = append(index.columns, column)
 			}
+		}
+	}
+
+	for _, column := range index.columns {
+		if !checkAddDropColumn(ctx, column) {
+			return nil
 		}
 	}
 
@@ -1255,14 +1281,17 @@ func (c *testCase) generateRenameIndex(repeat int) error {
 	return nil
 }
 
-func (c *testCase) prepareRenameIndex(_ interface{}, taskCh chan *ddlJobTask) error {
-	table := c.pickupRandomTable()
+func (c *testCase) prepareRenameIndex(ctx interface{}, taskCh chan *ddlJobTask) error {
+	table := c.getTable(ctx)
 	if table == nil || len(table.indexes) == 0 {
 		return nil
 	}
 	loc := rand.Intn(len(table.indexes))
 	index := table.indexes[loc]
 	newIndex := uuid.NewV4().String()
+	if !checkModifyIdx(ctx, index) {
+		return nil
+	}
 	sql := fmt.Sprintf("ALTER TABLE `%s` RENAME INDEX `%s` to `%s`",
 		table.name, index.name, newIndex)
 	task := &ddlJobTask{
@@ -1281,6 +1310,7 @@ func (c *testCase) renameIndexJob(task *ddlJobTask) error {
 	if c.isTableDeleted(table) {
 		return fmt.Errorf("table %s is not exists", table.name)
 	}
+
 	iOfRenameIndex := -1
 	for i := range table.indexes {
 		if arg.index.name == table.indexes[i].name {
@@ -1306,8 +1336,8 @@ func (c *testCase) generateDropIndex(repeat int) error {
 	return nil
 }
 
-func (c *testCase) prepareDropIndex(_ interface{}, taskCh chan *ddlJobTask) error {
-	table := c.pickupRandomTable()
+func (c *testCase) prepareDropIndex(ctx interface{}, taskCh chan *ddlJobTask) error {
+	table := c.getTable(ctx)
 	if table == nil {
 		return nil
 	}
@@ -1316,6 +1346,9 @@ func (c *testCase) prepareDropIndex(_ interface{}, taskCh chan *ddlJobTask) erro
 	}
 	indexToDropIndex := rand.Intn(len(table.indexes))
 	indexToDrop := table.indexes[indexToDropIndex]
+	if !checkModifyIdx(ctx, indexToDrop) {
+		return nil
+	}
 	sql := fmt.Sprintf("ALTER TABLE `%s` DROP INDEX `%s`", table.name, indexToDrop.name)
 
 	arg := &ddlIndexJobArg{index: indexToDrop}
@@ -1388,13 +1421,16 @@ func (c *testCase) generateAddColumn(repeat int) error {
 	return nil
 }
 
-func (c *testCase) prepareAddColumn(_ interface{}, taskCh chan *ddlJobTask) error {
-	table := c.pickupRandomTable()
+func (c *testCase) prepareAddColumn(ctx interface{}, taskCh chan *ddlJobTask) error {
+	table := c.getTable(ctx)
 	if table == nil {
 		return nil
 	}
 	strategy := rand.Intn(ddlTestAddDropColumnStrategyAtRandom) + ddlTestAddDropColumnStrategyAtBeginning
 	newColumn := getRandDDLTestColumn()
+	if !checkAddDropColumn(ctx, newColumn) {
+		return nil
+	}
 	insertAfterPosition := -1
 	// build SQL
 	sql := fmt.Sprintf("ALTER TABLE `%s` ADD COLUMN `%s` %s", table.name, newColumn.name, newColumn.getDefinition())
@@ -1406,6 +1442,9 @@ func (c *testCase) prepareAddColumn(_ interface{}, taskCh chan *ddlJobTask) erro
 	case ddlTestAddDropColumnStrategyAtRandom:
 		insertAfterPosition = rand.Intn(table.columns.Size())
 		column := getColumnFromArrayList(table.columns, insertAfterPosition)
+		if !checkRelatedColumn(ctx, column) {
+			return nil
+		}
 		sql += fmt.Sprintf(" AFTER `%s`", column.name)
 	}
 
@@ -1472,8 +1511,8 @@ func (c *testCase) generateModifyColumn2(repeat int) error {
 	return nil
 }
 
-func (c *testCase) prepareModifyColumn2(_ interface{}, taskCh chan *ddlJobTask) error {
-	table := c.pickupRandomTable()
+func (c *testCase) prepareModifyColumn2(ctx interface{}, taskCh chan *ddlJobTask) error {
+	table := c.getTable(ctx)
 	if table == nil {
 		return nil
 	}
@@ -1494,10 +1533,16 @@ func (c *testCase) prepareModifyColumn2(_ interface{}, taskCh chan *ddlJobTask) 
 		}
 		// for change column
 		modifiedColumn = generateRandModifiedColumn2(origColumn, true)
+		if !checkAddDropColumn(ctx, origColumn) || !checkAddDropColumn(ctx, modifiedColumn) {
+			return nil
+		}
 		origColumn.setRenamed()
 		sql = fmt.Sprintf("alter table `%s` change column `%s` `%s` %s", table.name,
 			origColumn.name, modifiedColumn.name, modifiedColumn.getDefinition())
 	} else {
+		if !checkModifyColumn(ctx, origColumn) {
+			return nil
+		}
 		// for modify column
 		modifiedColumn = generateRandModifiedColumn2(origColumn, false)
 		sql = fmt.Sprintf("alter table `%s` modify column `%s` %s", table.name,
@@ -1511,12 +1556,20 @@ func (c *testCase) prepareModifyColumn2(_ interface{}, taskCh chan *ddlJobTask) 
 		sql += " FIRST"
 	case ddlTestAddDropColumnStrategyAtEnd:
 		endColumn := getColumnFromArrayList(table.columns, table.columns.Size()-1)
+		if !checkRelatedColumn(ctx, endColumn) {
+			origColumn.delRenamed()
+			return nil
+		}
 		if endColumn.name != origColumn.name {
 			sql += fmt.Sprintf(" AFTER `%s`", endColumn.name)
 		}
 	case ddlTestAddDropColumnStrategyAtRandom:
 		insertPosition := rand.Intn(table.columns.Size())
 		insertAfterColumn = getColumnFromArrayList(table.columns, insertPosition)
+		if !checkRelatedColumn(ctx, insertAfterColumn) {
+			origColumn.delRenamed()
+			return nil
+		}
 		if insertPosition != origColIndex {
 			sql += fmt.Sprintf(" AFTER `%s`", insertAfterColumn.name)
 		}
@@ -1548,8 +1601,8 @@ func (c *testCase) generateModifyColumn(repeat int) error {
 	return nil
 }
 
-func (c *testCase) prepareModifyColumn(_ interface{}, taskCh chan *ddlJobTask) error {
-	table := c.pickupRandomTable()
+func (c *testCase) prepareModifyColumn(ctx interface{}, taskCh chan *ddlJobTask) error {
+	table := c.getTable(ctx)
 	if table == nil {
 		return nil
 	}
@@ -1567,11 +1620,17 @@ func (c *testCase) prepareModifyColumn(_ interface{}, taskCh chan *ddlJobTask) e
 			return nil
 		}
 		modifiedColumn = generateRandModifiedColumn(origColumn, true)
+		if !checkAddDropColumn(ctx, origColumn) || !checkAddDropColumn(ctx, modifiedColumn) {
+			return nil
+		}
 		origColumn.setRenamed()
 		sql = fmt.Sprintf("alter table `%s` change column `%s` `%s` %s", table.name,
 			origColumn.name, modifiedColumn.name, modifiedColumn.getDefinition())
 	} else {
 		modifiedColumn = generateRandModifiedColumn(origColumn, false)
+		if !checkModifyColumn(ctx, origColumn) {
+			return nil
+		}
 		sql = fmt.Sprintf("alter table `%s` modify column `%s` %s", table.name,
 			origColumn.name, modifiedColumn.getDefinition())
 	}
@@ -1582,12 +1641,20 @@ func (c *testCase) prepareModifyColumn(_ interface{}, taskCh chan *ddlJobTask) e
 		sql += " FIRST"
 	case ddlTestAddDropColumnStrategyAtEnd:
 		endColumn := getColumnFromArrayList(table.columns, table.columns.Size()-1)
+		if !checkRelatedColumn(ctx, endColumn) {
+			origColumn.delRenamed()
+			return nil
+		}
 		if endColumn.name != origColumn.name {
 			sql += fmt.Sprintf(" AFTER `%s`", endColumn.name)
 		}
 	case ddlTestAddDropColumnStrategyAtRandom:
 		insertPosition := rand.Intn(table.columns.Size())
 		insertAfterColumn = getColumnFromArrayList(table.columns, insertPosition)
+		if !checkRelatedColumn(ctx, insertAfterColumn) {
+			origColumn.delRenamed()
+			return nil
+		}
 		if insertPosition != origColIndex {
 			sql += fmt.Sprintf(" AFTER `%s`", insertAfterColumn.name)
 		}
@@ -1673,8 +1740,8 @@ func (c *testCase) generateDropColumn(repeat int) error {
 	return nil
 }
 
-func (c *testCase) prepareDropColumn(_ interface{}, taskCh chan *ddlJobTask) error {
-	table := c.pickupRandomTable()
+func (c *testCase) prepareDropColumn(ctx interface{}, taskCh chan *ddlJobTask) error {
+	table := c.getTable(ctx)
 	if table == nil {
 		return nil
 	}
@@ -1696,6 +1763,10 @@ func (c *testCase) prepareDropColumn(_ interface{}, taskCh chan *ddlJobTask) err
 	}
 
 	columnToDrop := getColumnFromArrayList(table.columns, columnToDropIndex)
+
+	if !checkAddDropColumn(ctx, columnToDrop) {
+		return nil
+	}
 
 	// Primary key columns cannot be dropped
 	if columnToDrop.isPrimaryKey {
